+++ asdf_astropy/converters/transform/core.py
@@ -1,7 +1,5 @@
 import abc
-import warnings
 
-import numpy as np
 from asdf.extension import Converter
 
 from ..utils import import_type
@@ -125,56 +123,50 @@ class TransformConverterBase(Converter):
         return node
 
     def _serialize_bounding_box(self, model, node):
-        import astropy
-        from packaging.version import Version
-
-        if Version(astropy.__version__) > Version("4.999.999"):
-            self._serialize_bounding_box_astropy_5(model, node)
-        else:
-            self._serialize_bounding_box_astropy_4(model, node)
-
-    def _serialize_bounding_box_astropy_4(self, model, node):
-        try:
-            bb = model.bounding_box
-        except NotImplementedError:
-            bb = None
-
-        if bb is not None:
-            if model.n_inputs == 1:
-                bb = list(bb)
-            else:
-                bb = [list(item) for item in model.bounding_box]
-            node["bounding_box"] = bb
-
-    def _serialize_bounding_box_astropy_5(self, model, node):
         from astropy.modeling.bounding_box import CompoundBoundingBox, ModelBoundingBox
 
-        try:
-            bb = model.bounding_box
-        except NotImplementedError:
-            bb = None
+        # ignore any default bounding_box
+        if (bbox := model._user_bounding_box) is not None:
+            if isinstance(bbox, ModelBoundingBox):
+                self._serialize_bbox(model, node)
+            elif isinstance(bbox, CompoundBoundingBox):
+                self._serialize_cbbox(model, node)
 
-        if isinstance(bb, ModelBoundingBox):
-            bb = bb.bounding_box(order="C")
+    def _serialize_bbox(self, model, node):
+        from astropy.modeling.bounding_box import ModelBoundingBox
+        from astropy.utils import minversion
+
+        bbox = model.bounding_box
+
+        if minversion("asdf_transform_schemas", "0.2.2", inclusive=False):
+            if len(bbox.ignored) > 0:
+                if minversion("astropy", "5.1"):
+                    kwargs = {"_preserve_ignore": True}
+                else:
+                    raise RuntimeError("Bounding box ignored arguments are only supported by astropy 5.1+")
+            else:
+                kwargs = {}
+
+            bbox = ModelBoundingBox.validate(model, bbox, **kwargs)
+        else:
+            bbox = bbox.bounding_box(order="C")
 
             if model.n_inputs == 1:
-                bb = list(bb)
+                bbox = list(bbox)
             else:
-                bb = [list(item) for item in bb]
-            node["bounding_box"] = bb
+                bbox = [list(item) for item in bbox]
 
-        elif isinstance(bb, CompoundBoundingBox):
-            selector_args = [[sa.index, sa.ignore] for sa in bb.selector_args]
-            node["selector_args"] = selector_args
-            node["cbbox_keys"] = list(bb.bounding_boxes.keys())
+        node["bounding_box"] = bbox
 
-            bounding_boxes = list(bb.bounding_boxes.values())
-            if len(model.inputs) - len(selector_args) == 1:
-                node["cbbox_values"] = [list(sbbox.bounding_box()) for sbbox in bounding_boxes]
-            else:
-                node["cbbox_values"] = [
-                    [list(item) for item in sbbox.bounding_box() if np.isfinite(item[0])] for sbbox in bounding_boxes
-                ]
+    def _serialize_cbbox(self, model, node):
+        from astropy.utils import minversion
+
+        bbox = model.bounding_box
+
+        if minversion("asdf_transform_schemas", "0.2.2", inclusive=False):
+            node["bounding_box"] = bbox
+        else:
+            raise RuntimeError("Need a schema in order to serialize a compound bounding_box")
 
     def from_yaml_tree(self, node, tag, ctx):
         from astropy.modeling.core import CompoundModel
@@ -184,13 +176,13 @@ class TransformConverterBase(Converter):
         if "name" in node:
             model.name = node["name"]
 
-        self._deserialize_bounding_box(model, node)
-
         if "inputs" in node:
             model.inputs = tuple(node["inputs"])
 
         if "outputs" in node:
             model.outputs = tuple(node["outputs"])
+
+        self._deserialize_bounding_box(model, node)
 
         param_and_model_constraints = {}
         for constraint in ["fixed", "bounds"]:
@@ -209,31 +201,15 @@ class TransformConverterBase(Converter):
             model.inverse = node["inverse"]
 
     def _deserialize_bounding_box(self, model, node):
-        import astropy
-        from packaging.version import Version
-
-        if Version(astropy.__version__) > Version("4.999.999"):
-            self._deserialize_bounding_box_astropy_5(model, node)
-        else:
-            self._deserialize_bounding_box_astropy_4(model, node)
-
-    def _deserialize_bounding_box_astropy_4(self, model, node):
         if "bounding_box" in node:
-            model.bounding_box = node["bounding_box"]
-        elif "selector_args" in node:
-            warnings.warn("This version of astropy does not support compound bounding boxes.")
+            bounding_box = node["bounding_box"]
 
-    def _deserialize_bounding_box_astropy_5(self, model, node):
-        from astropy.modeling.bounding_box import CompoundBoundingBox
-
-        if "bounding_box" in node:
-            model.bounding_box = node["bounding_box"]
-        elif "selector_args" in node:
-            cbbox_keys = [tuple(key) for key in node["cbbox_keys"]]
-            bbox_dict = dict(zip(cbbox_keys, node["cbbox_values"]))
-
-            selector_args = node["selector_args"]
-            model.bounding_box = CompoundBoundingBox.validate(model, bbox_dict, selector_args)
+            if isinstance(bounding_box, list):
+                model.bounding_box = bounding_box
+            elif callable(bounding_box):
+                model.bounding_box = bounding_box(model)
+            else:
+                raise TypeError(f"Cannot form bounding_box from: {bounding_box}")
 
 
 class SimpleTransformConverter(TransformConverterBase):